Fix problems in review
