+++ .gitignore
@@ -34,6 +34,3 @@
 *.zim
 pip-selfcheck.json
 build/
-
-# IDE Settings
-.vscode/*
