+++ src/meson.build
@@ -36,8 +36,8 @@ executable('zimrecreate', ['zimrecreate.cpp', 'tools.cpp'],
 
 subdir('zimcheck')
 
-if with_writer
+if with_writer_and_mount
   subdir('zimwriterfs')
+  subdir('zimmount')
 endif
 
-subdir('zimmount')
